Added script to delete files in Unix/Linux OS
The script is to define a config/text file with a list of files to be deleted. Then the script will read all the files listed in the config/text and deleted them. The files to be deleted must be defined as an array in the config/text file.

I.E.:

cat << EOF >> list_of_files_to_delete.txt
FILES_TO_DELETE=(
"file1"
"file2"
"file3"
"path\file4"
)